dom/browser-element/mochitest/test_browserFrame10.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=702880
-->
 <head>
  <title>Test for Bug 702880</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserFrameHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=753595">Mozilla Bug 702880</a>

<!--
   Test the setVisible property for mozbrowser
-->

<script type="application/javascript;version=1.7">

SimpleTest.waitForExplicitFinish();

var iframeScript = function() {
  content.document.addEventListener("mozvisibilitychange", function() {
    sendAsyncMessage('test:visibilitychange', {
      hidden: content.document.mozHidden
    });
  }, false);
}

function runTest() {

  browserFrameHelpers.setEnabledPref(true);
  browserFrameHelpers.addToWhitelist();

  var mm;
  var numEvents = 0;
  var iframe1 = document.createElement('iframe');
  iframe1.mozbrowser = true;
  iframe1.src = 'data:text/html,1';

  document.body.appendChild(iframe1);

  function recvVisibilityChanged(msg) {
    numEvents++;
    if (numEvents === 1) {
      ok(true, 'iframe recieved visibility changed');
      ok(msg.json.hidden === true, 'mozHidden attribute correctly set');
      iframe1.setVisible(false);
      iframe1.setVisible(true);
    } else if (numEvents === 2) {
      ok(msg.json.hidden === false, 'mozHidden attribute correctly set');
      // Allow some time in case we generate too many events
      setTimeout(function() {
        mm.removeMessageListener('test:visibilitychange', recvVisibilityChanged);
        SimpleTest.finish();
      }, 100);
    } else {
      ok(false, 'Too many mozhidden events');
    }
  }

  function iframeLoaded() {
    mm = SpecialPowers.getBrowserFrameMessageManager(iframe1);
    mm.addMessageListener('test:visibilitychange', recvVisibilityChanged);
    mm.loadFrameScript('data:,(' + iframeScript.toString() + ')();', false);
    iframe1.setVisible(false);
  }

  iframe1.addEventListener('mozbrowserloadend', iframeLoaded);
}

addEventListener('load', function() { SimpleTest.executeSoon(runTest); });

</script>

</body>
</html>
